Hooking up deserialised output and fixing model data types

+++ src/Models/Configuration/WeatherConfiguration.cs
@@ -3,8 +3,8 @@ namespace Models.Configuration
     public class WeatherConfiguration
     {
         public string ActionUri { get; set; }
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public string Units { get; set; }
         public string ApiKey { get; set; }
     }
